% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthographic_distance.R
\name{check_epithets_match}
\alias{check_epithets_match}
\title{Check if Epithets match.}
\usage{
check_epithets_match(dist, epithets = NULL)
}
\arguments{
\item{dist}{data.frame with orthographic distances as returned from}

\item{epithets}{(optional) data.frame with epithets. Is NULL, row numbers of
matching epithets are returned. \code{orthographic_distance()}.}
}
\value{
a data.frame with epithets, distance, and if can be considered
  matching. If epithets = NULL, this returns only a vector with the row
  numbers of matchin epithets.
}
\description{
Check if the epithets can considered to be matching each others
according to Meyers et al. (2016). See also Jin & Yang (2020).
}
\examples{
x <- sapply(1:5000, function(x) paste(sample(letters[1:10], 5), collapse = ""))
y <- sapply(1:5000, function(x) paste(sample(letters[1:10], 5), collapse = ""))
dist <- orthographic_dist(x, y, parallel = TRUE)
matches <- check_epithets_match(dist, data.frame(x, y))
matches[matches$matching, ]
}
