% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{find_gbif}
\alias{find_gbif}
\title{Get GBIF taxonomy
Retrieve GBIF taxonomy using the rgbif package.}
\usage{
find_gbif(sp, parallel = FALSE, n_cores = 6, alternative = FALSE)
}
\arguments{
\item{sp}{species name.}

\item{parallel}{logical, weather to parallelize or not.}

\item{n_cores}{number of cores to use in parallilization.}

\item{alternative}{logical; TRUE return alternative instead of main result.}
}
\value{
a string with the species name taxonomy suggested by GBIF.
}
\description{
Get GBIF taxonomy
Retrieve GBIF taxonomy using the rgbif package.
}
\details{
Calls sequentially \code{rm_whitespace()}, \code{rm_extra()},
\code{rm_abbr()}, \code{rm_punct()}, \code{rm_whitespace()},
\code{rm_cf()}, and \code{rm_var()}.
}
