% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{fix_typos}
\alias{fix_typos}
\title{Fix typos
programmatically fix typos using \code{taxize} R package. First, remove typos
using \code{rm_typos()}, then check the Catalogue Of Life (COL) using taxize
to find possible matches of erroneous species names.}
\usage{
fix_typos(sp, parallel = FALSE, n_cores = 6)
}
\arguments{
\item{sp}{species name.}

\item{parallel}{logical, weather to parallelize or not.}

\item{n_cores}{number of cores to use in parallilization.}
}
\value{
a string with the species name taxonomy suggested by COL.
}
\description{
Fix typos
programmatically fix typos using \code{taxize} R package. First, remove typos
using \code{rm_typos()}, then check the Catalogue Of Life (COL) using taxize
to find possible matches of erroneous species names.
}
\details{
Species not matched are returned as 'NA'; species without
species-level taxonomy are returned as 'Species NA'
}
