% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{gbif_table}
\alias{gbif_table}
\title{make table from GBIF data}
\usage{
gbif_table(df)
}
\arguments{
\item{df}{GBIF data.frame}
}
\value{
a data.frame table with taxonomic information
}
\description{
make table from GBIF data
}
