% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthographic_distance.R
\name{orthographic_dist}
\alias{orthographic_dist}
\title{Compute Absolute and Relative Orthographic Distance}
\usage{
orthographic_dist(x, y, parallel = FALSE, n_cores = 4)
}
\arguments{
\item{x}{first string or vector of strings.}

\item{y}{second string or vector of strings.}

\item{parallel}{TRUE/FALSE if to run in parallel.}

\item{n_cores}{number of cores for parallel computation.}
}
\value{
a data.frame with absolute and relative orthographic distances.
}
\description{
This function computes the absolute and relative orthographic
distance as discussed in Meyer et al. (2016). Absolute orthographic distance
is also known as Levenshtein distance, informally defined as the number of
corrections necessary to reach a string from another. Relative orthographic
distance is the absolute distance divided by the length of the string. Here,
the relative distance is obtained as the Levenshtein distance divided by the
average length of the strings.
}
