% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_abbr}
\alias{rm_abbr}
\title{Remove abbreviations
Remove abbreviation 'sp' from the species name.}
\usage{
rm_abbr(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without 'sp'.
}
\description{
Remove abbreviations
Remove abbreviation 'sp' from the species name.
}
