% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_cf}
\alias{rm_cf}
\title{Remove 'cf'
Remove 'cf' from the species name.}
\usage{
rm_cf(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without 'cf'.
}
\description{
Remove 'cf'
Remove 'cf' from the species name.
}
