% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_extra}
\alias{rm_extra}
\title{Remove additional information
Remove extra information not needed for taxonomic purposes (e.g. sex).}
\usage{
rm_extra(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without extra information.
}
\description{
Remove additional information
Remove extra information not needed for taxonomic purposes (e.g. sex).
}
