% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_punct}
\alias{rm_punct}
\title{Remove punctuation
Remove punctuation from the species name.}
\usage{
rm_punct(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without punctuation.
}
\description{
Remove punctuation
Remove punctuation from the species name.
}
