% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_typos}
\alias{rm_typos}
\title{Remove typos
Remove all unnecessary characters and punctuation from the species name.}
\usage{
rm_typos(sp, parallel = FALSE, n_cores = 6)
}
\arguments{
\item{sp}{species name.}

\item{parallel}{logical, weather to parallelize or not.}

\item{n_cores}{number of cores to use in parallilization.}
}
\value{
a string with the species name without typos.
}
\description{
Remove typos
Remove all unnecessary characters and punctuation from the species name.
}
\details{
Calls sequentially \code{rm_whitespace()}, \code{rm_extra()},
\code{rm_abbr()}, \code{rm_punct()}, \code{rm_whitespace()},
\code{rm_cf()}, and \code{rm_var()}.
}
