% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_var}
\alias{rm_var}
\title{remove subspecies
Remove subspecies and varieties}
\usage{
rm_var(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without subspecies and varieties
}
\description{
remove subspecies
Remove subspecies and varieties
}
