% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxclean.R
\name{rm_whitespace}
\alias{rm_whitespace}
\title{Remove whitespaces
Remove unnecessary whitespaces (e.g. two subsequent whitespaces)
from the species name.}
\usage{
rm_whitespace(sp)
}
\arguments{
\item{sp}{species name.}
}
\value{
a string with the species name without unnecessary whitespaces.
}
\description{
Remove whitespaces
Remove unnecessary whitespaces (e.g. two subsequent whitespaces)
from the species name.
}
