% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{write_pipeline}
\alias{write_pipeline}
\title{Write R script Pipeline}
\usage{
write_pipeline(
  pkgs = c("rgbif", "rgnparser", "RTNRS"),
  pipeline_script = "pipelines/template.R",
  write_intermediate_steps = FALSE
)
}
\arguments{
\item{pkgs}{packages to call sequentially.}

\item{pipeline_script}{file where to write the R code.}
}
\description{
Writes to a file the R code to reproduce a pipeline..
}
